Split toggleable monkeys and non-toggleable monkeys, update locale
