+++ MonkeyLoader.Resonite.Integration/Locale/en.json
@@ -41,8 +41,12 @@
     "Settings.MonkeyToggles.Breadcrumb": "Monkeys",
 
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Name": "Monkeys",
+    "MonkeyLoader.GamePacks.Resonite.MonkeysToggleable.Name": "Toggleable Monkeys",
+    "MonkeyLoader.GamePacks.Resonite.MonkeysNonToggleable.Name": "Non-Toggleable Monkeys",
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Description": "Lists the details of all available Monkeys of this Mod and allows toggling them.",
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Count.Name": "Count",
+    "MonkeyLoader.GamePacks.Resonite.MonkeysToggleable.Count.Name": "Count",
+    "MonkeyLoader.GamePacks.Resonite.MonkeysNonToggleable.Count.Name": "Count",
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Count.Description": "The number of Monkeys of this Mod.",
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Enabled.Name": "Enabled",
     "MonkeyLoader.GamePacks.Resonite.Monkeys.Enabled.Description": "Toggles whether this Monkey is active.",
